feat(graph): give the mirror projections an invoker (BI-FEDFABF6)
Three graph-mirror projections had no caller. Their only reference in the repository
was their own package.json definition, so the domains they maintain were empty while
the code graph — which HAS an indexer — stayed full.

MEASURED ON A FRESHLY PROVISIONED INSTALL

  code (CodeSymbol/CodeFile/...)   38,592 nodes   has an indexer
  EA / ArchiMate                      607 nodes   has an indexer
  knowledge (Wiki__*)                   0 nodes   NO invoker
  portfolio/DigitalProduct/Taxonomy     0 nodes   NO invoker
  DocPage + IMPACTS                     0 / 0     NO invoker

Every domain with a caller is populated; every domain without one is empty. That is
the same defect BI-3045CC18 ("Portfolio: 1" beside 16,147 code nodes) and BI-0E019B95
(zero DocPage against a committed 616-edge manifest) each reported separately, now
reproduced from scratch on a clean install.

None of it fails or logs. The explorer renders a confident wrong answer — "nothing
documents this route", "this platform has one portfolio" — which is precisely the
failure mode these surfaces exist to prevent.

WHY BOOT

Boot follows migrations, follows every self-upgrade, and runs on first start of a new
install. Those are exactly the moments the source data moves relative to the mirror. A
schedule would also work but adds a cadence to tune; boot needs none and covers the
fresh-install case a schedule misses entirely.

Skipped under measurement runtime for the same reason the self-heal block beside it is:
a sweep portal measures a frozen baseline, and background writers racing the crawl are
the nondeterminism BI-232BA634 removed.

SAFETY PROPERTIES, EACH PINNED BY A TEST

- Isolated: one projection failing must not stop the others. A partially-current
  mirror beats one that stopped at the first error.
- Never throws: this runs inside instrumentation.register(). A stale mirror is bad; an
  unbootable portal is worse.
- Owned-scope only: doc-impact clears ONLY DocPage, the one label it owns. Clearing a
  shared label would delete code-graph nodes and their IMPORTS/TESTED_BY edges — the
  destruction BI-EC5FF1A0 fixed.

All four tests were verified to FAIL against the naive implementation (straight-line
awaits, no isolation): 2 failed / 2 passed naive, 4 passed with the isolation in place.

IMPORTING THE REBUILD MUST NOT RUN IT

rebuild-knowledge-and-portfolio-graph.ts executed main() at module scope and then
disconnected the shared Prisma client, so importing it from the portal would have run
a full rebuild as a side effect of an import and pulled the connection out from under
the caller. It now exports rebuildKnowledgeAndPortfolioGraph and guards the CLI entry
behind an argv check. Verified: importing the module yields the export and performs no
writes.

MODULE-SIZE BASELINE

apps/web/instrumentation.ts 1475 -> 1483 (+8, the boot hook). The ratchet's rule is
that a baselined file may only shrink, so this is a deliberate, visible exception
rather than something the guard waved through: the entry is edited by hand in the diff
so a reviewer can reject it, and `--update` was NOT used because it retightens every
unrelated file that happens to have shrunk. If the reviewer prefers, the alternative is
a scheduled task — but agent-task-scheduler.ts is already 808 LOC, over the soft
ceiling for an unbaselined file, so it trades one ratchet problem for another.

A GUARD CAUGHT A REAL REUSE VIOLATION

The first draft carried a local `error instanceof Error ? error.message : String(error)`
helper. check-no-raw-error-message flagged it; it now uses the shared getErrorMessage.

VERIFICATION

52 repo guards clean, 92 tests across lib/graph, web typecheck 0 errors. The two
projections were also run against this install to repair it: portfolio 0 -> 765,
knowledge 0 -> 354 wiki + 198 doc, IMPACTS 0 -> 732, and the route-to-doc traversal
resolves (/admin/storefront -> index).

STILL OPEN

BI-A73954F7 — nothing reports projection freshness, so an empty or destroyed domain is
still indistinguishable from a true answer. That is the observability step and belongs
after this one; a drift guard belongs after both.

Docs-Impact-Decision: no-user-facing-route-change — boot-time data reconciliation; no
route, control name, or documented workflow changes.
